fix DO_NOT_LAUNCH.firm link in italian version of finalizing setup
yes this isnt crowdin but bleh. the issue is caused by the link pointing to /assets/do_not_launch.firm. on the english version, this points to 3ds.hacks.guide/assets/do_not_launch.firm which is valid. on translations, this point to 3ds.hacks.guide/[language]/assets/do_not_launch.firm which is not valid. i imagine there's a more proper fix for this but i reckon this will work for now