Remove reference to Nintendo 3DS for finalize.romfs placement.
While Nintendo 3DS is probably the folder they'll end up putting it in (many people confuse that with the root), there are other folders that the script can detect. They really don't need to know where they misplaced it, just that they did, and they should enter the keycombo to fix it.