Add a specific include for redirecting to the wiki
This will add a notice that users will be redirected to an external
site; while the Hacks Guide Wiki is technically still us, it's still
good practice and less of a shock to users when it goes to somewhere
that looks and feels totally different.

The page will automatically redirect after 5 seconds.