installing-boot9strap-(mset9-cli): recommend against Microsoft Store
The version of Python from the Microsoft Store does not work with MSET9,
in part because this version doesn't include the "py" launcher, which is
what MSET9-Windows.bat wants. I don't know if it does actually work if
mset9.py is used directly though, but until then, the python.org
installers should be recommended.