Bring back cookieconsent
It was removed during the transition to VitePress. We do need to add it
back given the structure of the site.

The old plugin is back, but with some changes:
- instead of static banner at top, it is a floating box on the bottom
  right
- a custom CSS override was added to adapt the theming of the box to
  VitePress
- make translatable by adding a locale-specific head that imports the
  JS file with a hack to add custom parameters for the message content
- minor rice

For now this is only added to the English page. As the translations for
it start to fill, we can enable for remaining languages.